GPSLC.jl Documentation

GPSLC.samplePosterior — Function

Draw samples from the posterior given the observed data X,T,Y.

Params:

X: Input covariates

T: Input treatment

Y: Output

SigmaU: Object structure

hyperParams=getHyperParameters(): Hyperparameters for posterior sampling

Returns:

posteriorSample: samples from posterior determined by hyperparameters

GPSLC.getHyperParameters — Function

Hyperparameters

Defaults are those used in original paper, listed here for modification

uNoiseShape::Float64=4.0: shape parameter of the InvGamma prior over the noise of U

uNoiseScale::Float64=4.0: scale parameter of the InvGamma prior over the noise of U

xNoiseShape::Float64=4.0: shape parameter of the InvGamma prior over the noise of X

xNoiseScale::Float64=4.0: scale parameter of the InvGamma prior over the noise of X

tNoiseShape::Float64=4.0: shape parameter of the InvGamma prior over the noise of T

tNoiseScale::Float64=4.0: scale parameter of the InvGamma prior over the noise of T

yNoiseShape::Float64=4.0: shape parameter of the InvGamma prior over the noise of Y

yNoiseScale::Float64=4.0: scale parameter of the InvGamma prior over the noise of Y

xScaleShape::Float64=4.0: shape parameter of the InvGamma prior over kernel scale of X

xScaleScale::Float64=4.0: scale parameter of the InvGamma prior over kernel scale of X

tScaleShape::Float64=4.0: shape parameter of the InvGamma prior over kernel scale of T

tScaleScale::Float64=4.0: scale parameter of the InvGamma prior over kernel scale of T

yScaleShape::Float64=4.0: shape parameter of the InvGamma prior over kernel scale of Y

yScaleScale::Float64=4.0: scale parameter of the InvGamma prior over kernel scale of Y

uxLSShape::Float64=4.0: shape parameter of the InvGamma prior over kernel lengthscale of U and X

uxLSScale::Float64=4.0: scale parameter of the InvGamma prior over kernel lengthscale of U and X

utLSShape::Float64=4.0: shape parameter of the InvGamma prior over kernel lengthscale of U and T

utLSScale::Float64=4.0: scale parameter of the InvGamma prior over kernel lengthscale of U and T

xtLSShape::Float64=4.0: shape parameter of the InvGamma prior over kernel lengthscale of X and T

xtLSScale::Float64=4.0: scale parameter of the InvGamma prior over kernel lengthscale of X and T

uyLSShape::Float64=4.0: shape parameter of the InvGamma prior over kernel lengthscale of U and Y

uyLSScale::Float64=4.0: scale parameter of the InvGamma prior over kernel lengthscale of U and Y

xyLSShape::Float64=4.0: shape parameter of the InvGamma prior over kernel lengthscale of X and Y

xyLSScale::Float64=4.0: scale parameter of the InvGamma prior over kernel lengthscale of X and Y

tyLSShape::Float64=4.0: shape parameter of the InvGamma prior over kernel lengthscale of T and Y

tyLSScale::Float64=4.0: scale parameter of the InvGamma prior over kernel lengthscale of T and Y

GPSLC.sampleITE — Function

Estimate Individual Treatment Effect using GPSLC model

Params:

X: Input covariates

T: Input treatment

Y: Output

SigmaU: Object structure

posteriorSample=samplePosterior(X,T,Y,SigmaU): Samples of parameters from posterior

Returns:

ITEsamples: n x m matrix where n is the number of data, and m is the number of samples

GPSLC.summarizeITE — Function

Create dataframe of mean, lower and upper quantiles of the ITE samples.

Params:

ITEsamples: n x m array of ITE samples

savetofile: Optionally save the resultant dataframe as CSV to the filename passed.

Returns:

df: Dataframe of Individual, Mean, LowerBound, and UpperBound values for the samples.